+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 _Breaking Changes_
 * Add jest config to `__tests__` directories automatically
 * `object-curly-spacing` changed from `[2, 'never']` to `[2, 'always']`
+* Require eslint@^6.8.0
 
 ## 4.0.0
 _Breaking Changes_