+++ CHANGELOG.md
@@ -7,9 +7,9 @@
 _Breaking Changes_
 
 - Add jest config to `__tests__` directories automatically
-- `object-curly-spacing` changed from `[2, 'never']` to `[2, 'always']`
-- Require eslint@^6.8.0
+- Require eslint@^7.3.1
 - Use prettier for style rules
+- Use typescript for syntax rules
 
 ## 4.0.0