+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 # unreleased
 
 - Add typescript types
-- Dependency Updates
+- Dependency Updates:
+  - Node to v18
+- DevDependency Updates
   - typescript-eslint/eslint-plugin to v6
   - @typescript-eslint/parser to v6
   - typescript to v5